requirements.txt
pandas
tensorflow
keras
matplotlib
numpy==1.24.4
scipy==1.13.1
mlflow
joblib 
python-box
ensure
dotenv
tensorflow
pathlib
streamlit

